:construction: setup banking domain in operations
