+++ libs/elements/layout/page/src/lib/components/sidemenu/sidemenu.component.scss
@@ -124,3 +124,8 @@ a:hover {
   background-color: #fafafa;
   border-radius: 20px 0px 0px 20px;
 }
+
+
+.sub-menu {
+  margin-left: 20px;
+}
