+++ libs/features/finance/banking/activate-banking/src/lib/pages/banking-page/banking-page.component.scss
@@ -1,0 +1,11 @@
+.main-page {
+  padding: 50px;
+}
+
+.account { 
+  padding: 20px;
+}
+
+.account > div {
+  margin-bottom: 20px;
+}
