+++ libs/features/operations/main/src/lib/operations-main.module.ts
@@ -13,8 +13,6 @@ import { iTalPageModule } from '@app/elements/layout/page';
 
 import { OperationsPageComponent } from './pages/operations-page/operations-page.component';
 
-import { ActivatePontoBankingService } from './services/activate-ponto-banking.service';
-
 import { OperationsRouterModule } from './operations.router';
 
 
@@ -32,6 +30,6 @@ import { OperationsRouterModule } from './operations.router';
     OperationsRouterModule
   ],
   declarations: [OperationsPageComponent],
-  providers: [ActivatePontoBankingService]
+  providers: []
 })
 export class OperationsModule {}
