+++ libs/features/operations/main/src/lib/operations.router.ts
@@ -9,7 +9,6 @@ const OPERATIONS_ROUTES: Route[] = [
   {
     path: '',
     component: OperationsPageComponent,
-    
   },
   {
     path: 'banking',
