--- libs/features/operations/main/src/lib/pages/operations-page/operations-page.component.scss
@@ -1,11 +1,0 @@
-.main-page {
-  padding: 50px;
-}
-
-.account { 
-  padding: 20px;
-}
-
-.account > div {
-  margin-bottom: 20px;
-}
