+++ libs/features/operations/main/tsconfig.lib.json
@@ -13,5 +13,5 @@
     "jest.config.ts",
     "src/**/*.test.ts"
   ],
-  "include": ["src/**/*.ts"]
+  "include": ["src/**/*.ts", "../../finance/banking/activate-banking/src/lib/services/activate-ponto-banking.service.ts"]
 }
